[refactor] Moved main logic of dmm.ts
